+++ src/lib.rs
@@ -46,6 +46,9 @@ extern crate log;
 #[macro_use]
 extern crate lazy_static;
 
+#[macro_use]
+extern crate serde_derive;
+
 pub mod sink;
 pub mod buckets;
 pub mod config;
