--- src/metric/event_types.in.rs
@@ -1,6 +1,0 @@
-#[derive(PartialEq, Debug, Serialize, Deserialize, Clone)]
-pub enum Event {
-    Telemetry(sync::Arc<Option<Telemetry>>),
-    Log(sync::Arc<Option<LogLine>>),
-    TimerFlush,
-}
