--- src/metric/tagmap_types.in.rs
@@ -1,4 +1,0 @@
-#[derive(Clone,Debug,PartialEq,Serialize,Deserialize)]
-pub struct TagMap<K, V> {
-    inner: Vec<(K, V)>,
-}
