+++ src/sink/mod.rs
@@ -36,8 +36,9 @@ pub trait Sink {
     fn valve_state(&self) -> Valve;
     fn deliver(&mut self, point: sync::Arc<Option<Telemetry>>) -> ();
     fn deliver_line(&mut self, line: sync::Arc<Option<LogLine>>) -> ();
-    fn run(&mut self, mut recv: hopper::Receiver<Event>) {
+    fn run(&mut self, recv: hopper::Receiver<Event>) {
         let mut attempts = 0;
+        let mut recv = recv.into_iter();
         loop {
             time::delay(attempts);
             match recv.next() {
